+++ packages/web-components/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.141.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.140.0...@carbon-labs/web-components@0.141.0) (2026-04-30)
+
+**Note:** Version bump only for package @carbon-labs/web-components
+
+
+
+
+
+# [0.140.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.139.0...@carbon-labs/web-components@0.140.0) (2026-04-27)
+
+**Note:** Version bump only for package @carbon-labs/web-components
+
+
+
+
+
 # [0.139.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/web-components@0.138.0...@carbon-labs/web-components@0.139.0) (2026-04-22)
 
 **Note:** Version bump only for package @carbon-labs/web-components